+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.1] - 2022-07-04
+## [0.0.1] - 2022-08-15
 
 ### Added
 
@@ -16,6 +16,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Save-MispContext
 - Read-MispContext
 - Invoke-MispRestMethod
+- Get-MispEvent
+- Search-MispEvent
+- Get-MispAttribute
+- Search-Misp-Attribute
+- New-MispAttribute
+- New-MispEvent
+- Get-MispWarningList
+- Get-MispTag
+- Add-MispEventTag
+- Add-MispAttributeTag
 
 ### Changed
 
@@ -39,7 +49,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Known Issues
 
-- [#1](https://url.to.repo/repo/path/-/issues/1) Documentation is somewhat lacking
+- [#1](https://github.com/IPSecMSSP/misp.tools/issues/1) Documentation is somewhat lacking
 
-[Unreleased]: https://url.to.repo/repo/path/
-[0.0.1]: https://url.to.repo/repo/path/-/releases/v0.0.1
+[Unreleased]: https://github.com/IPSecMSSP/misp.tools
+[0.0.1]: https://github.com/IPSecMSSP/misp.tools/releases/v0.0.1